fix delete

+++ application/controllers/api/Admin.php
@@ -133,6 +133,12 @@ class Admin extends REST_Controller {
 	public function index_delete($id)
 	{
 		if ($this->checkToken() !== true) return $this->response($this->checkToken(), REST_Controller::HTTP_UNAUTHORIZED);
+
+		if ($this->Admin->getById($id) == null) return $this->response([
+			'success' => false,
+			'message' => "Data Tidak Ditemukan",
+		], REST_Controller::HTTP_NOT_FOUND);
+
 		if (!$this->Admin->delete($id)) {
 			return $this->response([
 				'success' => false,
