+++ application/controllers/api/Transaksi.php
@@ -147,6 +147,12 @@ class Transaksi extends REST_Controller
 	public function index_delete($id)
 	{
 		if ($this->checkToken() !== true) return $this->response($this->checkToken(), REST_Controller::HTTP_UNAUTHORIZED);
+
+		if ($this->Transaksi->getById($id) == null) return $this->response([
+			'success' => false,
+			'message' => "Data Tidak Ditemukan",
+		], REST_Controller::HTTP_NOT_FOUND);
+
 		if (!$this->Transaksi->delete($id)) {
 			return $this->response([
 				'success' => false,
